Write sudo code for program to pick a random word from standard input
Goal of program is to prompt user to enter any number of words to be
picked at random as the winner. Sudo code lays out method of using a
while loop to keep checking if standard input is empty or not. Once
standard input is empty, the while loop will stop and the program
will chose the winning word